Moved duplicate plugin code to single location, and small improvements
