+++ lib/api/alexa/index.js
@@ -1,10 +1,8 @@
 'use strict';
 
 var moment = require('moment');
-var _each = require('lodash/each');
 
 function configure (app, wares, ctx, env) {
-  var entries = ctx.entries;
   var express = require('express')
     , api = express.Router( );
   var translate = ctx.language.translate;
@@ -16,30 +14,7 @@ function configure (app, wares, ctx, env) {
   // json body types get handled as parsed json
   api.use(wares.bodyParser.json());
 
-  ctx.plugins.eachEnabledPlugin(function each(plugin){
-    if (plugin.virtAsst) {
-      if (plugin.virtAsst.intentHandlers) {
-        console.log('Alexa: Plugin ' + plugin.name + ' supports Virtual Assistants');
-        _each(plugin.virtAsst.intentHandlers, function (route) {
-          if (route) {
-            ctx.alexa.configureIntentHandler(route.intent, route.intentHandler, route.metrics);
-          }
-        });
-      }
-      if (plugin.virtAsst.rollupHandlers) {
-        console.log('Alexa: Plugin ' + plugin.name + ' supports rollups for Virtual Assistants');
-        _each(plugin.virtAsst.rollupHandlers, function (route) {
-          console.log('Route');
-          console.log(route);
-          if (route) {
-            ctx.alexa.addToRollup(route.rollupGroup, route.rollupHandler, route.rollupName);
-          }
-        });
-      }
-    } else {
-      console.log('Alexa: Plugin ' + plugin.name + ' does not support Virtual Assistants');
-    }
-  });
+  ctx.virtAsstBase.setupVirtAsstHandlers(ctx.alexa);
 
   api.post('/alexa', ctx.authorization.isPermitted('api:*:read'), function (req, res, next) {
     console.log('Incoming request from Alexa');
@@ -82,70 +57,7 @@ function configure (app, wares, ctx, env) {
     }
   });
 
-  ctx.alexa.addToRollup('Status', function bgRollupHandler(slots, sbx, callback) {
-    entries.list({count: 1}, function (err, records) {
-      var direction;
-      if (translate(records[0].direction)) {
-        direction = translate(records[0].direction);
-      } else {
-        direction = records[0].direction;
-      }
-      var status = translate('virtAsstStatus', {
-        params: [
-          sbx.scaleMgdl(records[0].sgv),
-          direction,
-          moment(records[0].date).from(moment(sbx.time))
-        ]
-      });
-      
-      callback(null, {results: status, priority: -1});
-    });
-  }, 'BG Status');
-
-  ctx.alexa.configureIntentHandler('MetricNow', function (callback, slots, sbx, locale) {
-    entries.list({count: 1}, function(err, records) {
-      var direction;
-      if(translate(records[0].direction)){
-        direction = translate(records[0].direction);
-      } else {
-        direction = records[0].direction;
-      }
-      var status = translate('virtAsstStatus', {
-        params: [
-          sbx.scaleMgdl(records[0].sgv),
-          direction,
-          moment(records[0].date).from(moment(sbx.time))]
-      });
-      
-      callback(translate('virtAsstTitleCurrentBG'), status);
-    });
-  }, ['bg', 'blood glucose', 'number']);
-
-  ctx.alexa.configureIntentHandler('MetricNow', function (callback, slots, sbx, locale) {
-    if (sbx.properties.delta && sbx.properties.delta.display) {
-      entries.list({count: 2}, function(err, records) {
-        callback(
-          translate('virtAsstTitleDelta'),
-          translate('virtAsstDelta', {
-            params: [
-              sbx.properties.delta.display == '+0' ? '0' : sbx.properties.delta.display,
-              moment(records[0].date).from(moment(sbx.time)),
-              moment(records[1].date).from(moment(sbx.time))
-            ]
-          })
-        );
-      });
-    } else {
-      callback(translate('virtAsstTitleDelta'), translate('virtAsstUnknown'));
-    }
-  }, ['delta']);
-
-  ctx.alexa.configureIntentHandler('NSStatus', function (callback, slots, sbx, locale) {
-    ctx.alexa.getRollup('Status', sbx, slots, locale, function (status) {
-      callback(translate('virtAsstTitleFullStatus'), status);
-    });
-  });
-
+  ctx.virtAsstBase.setupMutualIntents(ctx.alexa);
 
   function onLaunch(next) {
     console.log('Session launched');
@@ -181,6 +93,7 @@ function configure (app, wares, ctx, env) {
         metric = slots.metric.resolutions.resolutionsPerAuthority[0].values[0].value.name;
       } else {
         next(translate('virtAsstUnknownIntentTitle'), translate('virtAsstUnknownIntentText'));
+        return;
       }
     }
 
@@ -188,8 +101,10 @@ function configure (app, wares, ctx, env) {
     if (handler){
       var sbx = initializeSandbox();
       handler(next, slots, sbx);
+      return;
     } else {
       next(translate('virtAsstUnknownIntentTitle'), translate('virtAsstUnknownIntentText'));
+      return;
     }
   }
