+++ lib/server/bootevent.js
@@ -179,6 +179,10 @@ function boot (env, language) {
     ctx.dataloader = require('../data/dataloader')(env, ctx);
     ctx.notifications = require('../notifications')(env, ctx);
 
+    if (env.settings.isEnabled('alexa') || env.settings.isEnabled('googlehome')) {
+      ctx.virtAsstBase = require('../plugins/virtAsstBase')(env, ctx);
+    }
+
     if (env.settings.isEnabled('alexa')) {
       ctx.alexa = require('../plugins/alexa')(env, ctx);
     }
